+++ assets/assets.go
@@ -17,7 +17,6 @@ var (
 	ErrAssetInvalid       = errors.New("asset invalid")
 	ErrAssetDoesNotExist  = errors.New("asset does not exist")
 	ErrUnknownAssetSource = errors.New("unknown asset source")
-	ErrEthClientMissing   = errors.New("eth client is missing")
 )
 
 // TimeService ...
@@ -122,10 +121,6 @@ func (s *Service) assetFromDetails(assetID string, assetDetails *types.AssetDeta
 			builtin.New(assetID, assetDetails),
 		}, nil
 	case *types.AssetDetailsErc20:
-		if s.ethClient == nil {
-			// likely trying to do something with an ERC20 asset when using a null chain provider
-			return nil, ErrEthClientMissing
-		}
 		asset, err := erc20.New(assetID, assetDetails, s.nodeWallets.Ethereum, s.ethClient)
 		if err != nil {
 			return nil, err